Optimize kafka reporter
